feat:  Extract calendar component from DatePicker for reusability
