+++ packages/alto/src/components/DatePicker/Internal/index.ts
@@ -2,3 +2,4 @@ export { default as MonthsSwitcher } from './MonthsSwitcher.vue';
 export { default as Days } from './Days.vue';
 export { default as Presets } from './Presets.vue';
 export { default as YearsAndMonths } from './YearsAndMonths.vue';
+export { default as Calendar } from './Calendar.vue';
